don't propagate remote find on cluster render
Let's say you have two nodes in a cluster,
and they have metrics A, B.  You want to render "*",
and you do it from the node which has A.

http://Aserver/render/?target=*

A will first find all of the metrics to render (using find_all)
so it does the following:

http://Bserver/metrics/find/?local=1&format=pickle&query=*
and of course gets B

and gets A from it's own store.

So now we have pointers to A and B.  Time to render.

B is remote, so we call the remote pickle render on them
http://Bserver/render/?target=B&format=pickle

but, since local=1 isn't specified, it goes and asks around if
anyone has B (datalib.py)

http://Aserver/metrics/find/?local=1&format=pickle&query=B
of course, A doesn't have this, so it just renders its own B

hmmm, i think because of find_first there has to be an
escaped pattern