optionally use an LRU cache for the local store
if USE_LOCAL_STORAGE_CACHE is enabled (disabled by default)
the local store will make use of memoization when looking
for metrics which match a pattern.  This can be useful
if your disk I/O is limited and you want to save it for
the carbon cache, or if you are matching complex patterns

It keeps the most recent 1K entries.  The expiration of an
entry is configurable using the existing REMOTE_FIND_CACHE_DURATION
setting (default: 5 min)