Improve averageWeightedSeriesWithWildcards.
Previously this function took a string as its first argument
representing a pathExpression on which to substitute a regular
expression to map it to its companion "weight" series
pathExpression. Once the weightSeriesList was evaluated
we had to perform the same regular expression on the resulting
weightSeries paths, but the problem is that the regular expression
being applied to a pathExpression is different conceptually
than applying it to a path because a pathExpression may (and
commonly will) be a "glob" rather than a static path.

With this change the new argument handling expects a seriesList
for the average series argument, and we perform the regular
expression parsing on each path string to find its corresponding
weight path.

Also fixed the case where there is more than one average path
per weight path.